requirements.txt
streamlit
openai~=0.28.0
python-dotenv~=1.1.1
PyPDF2~=3.0.1
python-docx~=1.2.0


#cd StudyBuddyApp
#python3 -m venv venv
#source venv/bin/activate
#pip install -r requirements.txt
#streamlit run app.py


langchain~=0.3.26
dotenv~=0.9.9
langchain-openai~=0.3.27